agentcore-runtime-rag-agent-function/requirements.txt
mcp>=1.10.0
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
strands-agents-tools
strands-agents